refactor(engine): S2 Stage 5 — extract VST3 scanner + plugin-editor + missing-plugin domain to XlethSvcVst3 TU
Verbatim move of the 13 audio_* plugin handlers (scanner: scanPlugins,
getScanProgress, getScannedPlugins, getFailedPlugins, setMainWindowHandle;
editor windows: open/close/closeAll/isPluginEditorOpen; missing-plugin:
getMissingPlugins, retryMissingPlugin, removeAllMissing, resetCrashedPlugin)
out of XlethEngineService.cpp into engine/src/service/XlethSvcVst3.cpp/.h,
per docs/S2_SPLIT_PLAN.md §4 Stage 5.

getThisModuleDir promoted to external linkage — declared in
XlethSvcHelpers.h, definition stays in XlethEngineService.cpp since
Lifecycle's Initialize() is its sole external caller.

Note for future stages: audio/PluginRegistry.h (juce_audio_processors) must
be included before XlethSvcGlobals.h (which transitively pulls
juce_gui_extra via render/GridCompositor.h) — reversing that order breaks
JUCE's amalgamated module parsing.

Zero behavior change. ctest 43/47 (baseline, same 4 known failures); bridge
contract 12/15 (baseline, same 3 known failures); vitest 2504 passed/2
known-fail/3 skipped; app smoke 3/3; 7/7 end-to-end probe confirming
scanPlugins/getScanProgress/getMissingPlugins/setMainWindowHandle resolve
through the new TU and getThisModuleDir works from both call sites.
